Generate a User model with devise
This commit runs `rails g devise User role:string` to
generate a User model with a role.

I added a default for User role to be "user", this will
be the lowest level of access for the application, thus
making use of the deny by default principle.
I also altered the User table to include `first_name` and
`last_name` columns as I anticipate we will need this
later for personalised dashboard messages.

Finally I ran a `rails db:migrate`