Add components to root#index view
This adds personalised content to the index
view when users are logged in.

It also adds a sign in button if they are not
logged in, and a sign out button if they are.

This allows me to test whether the log in
feature is working by displaying the user's
first name and last name on the index view.