Add navigation bar to layout
This commit adds a conditionally rendered
navigation bar to the application layout.

Note the devise configuration was updated
to sign out via :get as this is an issue
with the gem. Sign out via :delete produces
a routing error. The same change has been made
to GOV.UK Signon to overcome the same issue.