Add a test for the feature
Adds a new feature test to check whether
a signed in user can view their remaining
annual leave on their dashboard.

Changes to the spec_helper.rb file are
needed in order to use the Devise helper
methods such as `sign_in` in RSpec feature
tests. They remove the deprecated
Devise::TestHelpers terminology and replace
it with up-to-date methods.