Add BankHolidayGenerator
This was taken from the govuk-frontend
application and adapted for the purposes
of this project:
- scotland and northern-ireland holidays
were removed from the BANK_HOLIDAYS hash
- altered the output of #perform to return
an array of Date objects
- removed all reference to title, notes and bunting
- removed all scotland and northern-ireland
holiday methods
- removed substitute arg from all methods
as this is just used to determine if a note
should be added to the output. Notes were
removed.

Link to original code:
https://github.com/alphagov/frontend/blob/51ed21f978560449c3fd339f56b5bd060245d3a1/lib/bank_holiday_generator.rb