Update helper method
I renamed the helper method for this because
it seemed more appropriate to name it
LineReportsHelper than UserHelper.

I've added a new method - #line_report_status
which returns the status of a line report's
annual leave requests, which is tested.

I then make use of the #line_report_status method
in the #line_reports_table_rows method so that it
returns correct information about a line report's
status when used in the line_reports#show view.
Previously this was hardcoded with a placeholder.
Now that users can make annual leave requests, it
is possible to update this field to show whether
the user has any pending annual leave requests.