Add annual_leave_request#update_status
Adds a route for patching the status of an annual
leave request with corresponding controller action.
This is used to update the status of an annual
leave request when approved, but will later also be
used for leave request denials.

Again I have scoped annual leave requests to only
those belonging to line reports managed by the signed
in line manager. This is to mitigate broken access
control.