Refactor _annual_leave_request_card partial
This partial was being used in almost the same form
in both line_reports#show and root#index, albeit with
different links on each card. I have refactored so
that both views use the same annual_leave_request_card
partial, but each has a separate links partial which is
passed through in the arguments. This eliminates a lot
of reused code.