Refactor email handling
I believe these changes will make it easier to extend
the `annual_leave_request#update_status` method to
accomodate other actions, such as editing and
withdrawing annual leave requests.

The present solution is very rigid, requiring new
EmailsHelper methods to be written for each new
update action. The new solution simply needs a new
StatusUpdate class to be added for the new action,
and for that action to be added to the case options
in the private `status_update` method in
AnnualLeaveRequestController. This brings the code much
closer to being open for extension.